refactor: export 방식 통일
