+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './use-boolean'
 export * from './use-interval'
 export * from './use-timer'
+export * from './use-counter'
