+++ src/layouts/main-layout.tsx
@@ -1,10 +1,10 @@
-import theme from '@/themes'
+import { theme } from '@/themes'
 import { createTheme, CssBaseline, ThemeProvider } from '@mui/material'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { useMemo } from 'react'
 import { Outlet } from 'react-router-dom'
 
-export default function MainLayout() {
+export function MainLayout() {
   const selectedTheme = createTheme(theme.light)
 
   const queryClient = useMemo(() => new QueryClient(), [])
