+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import '@/global.css'
 import { RouterProvider } from 'react-router-dom'
-import router from '@/routes'
+import { router } from '@/routes'
+import '@/global.css'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
