+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './join/join-page'
+export * from './login/login-page'
+export * from './main/main-page'
