+++ src/pages/join/components/index.ts
@@ -1,2 +1,4 @@
 export * from './join-stepper'
 export * from './term-form'
+export * from './details-form'
+export * from './info'
