--- src/pages/join/index.ts
@@ -1,3 +1,0 @@
-import { JoinPage } from './join-page'
-
-export default JoinPage
