--- src/pages/login/index.ts
@@ -1,3 +1,0 @@
-import { LoginPage } from './login-page'
-
-export default LoginPage
