--- src/pages/main/index.ts
@@ -1,3 +1,0 @@
-import { MainPage } from './main-page'
-
-export default MainPage
