+++ src/routes/paths.ts
@@ -1,7 +1,5 @@
-const paths = {
+export const paths = {
   root: '/',
   login: '/login',
   join: '/join',
 }
-
-export default paths
