+++ src/themes/index.ts
@@ -1,3 +1,1 @@
-import theme from './theme'
-
-export default theme
+export * from './theme'
